# laravel-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/comments": {
            "get": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Récupérer la liste des commentaires",
                "operationId": "getComments",
                "responses": {
                    "200": {
                        "description": "Liste des commentaires récupérée avec succès"
                    },
                    "500": {
                        "description": "Erreur interne"
                    }
                }
            },
            "post": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Créer un nouveau commentaire",
                "operationId": "createComment",
                "parameters": [
                    {
                        "name": "plan_id",
                        "in": "query",
                        "description": "ID du plan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "comment",
                        "in": "query",
                        "description": "Commentaire sur le plan",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Ceci est un commentaire."
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Commentaire créé avec succès"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/comments/{id}": {
            "get": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Récupérer un commentaire spécifique",
                "operationId": "getComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du commentaire",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commentaire récupéré avec succès"
                    },
                    "404": {
                        "description": "Commentaire non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Mettre à jour un commentaire",
                "operationId": "updateComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du commentaire",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plan_id",
                                    "comment"
                                ],
                                "properties": {
                                    "plan_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Mise à jour du commentaire."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Commentaire mis à jour avec succès"
                    },
                    "422": {
                        "description": "Données invalides"
                    },
                    "404": {
                        "description": "Commentaire non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Supprimer un commentaire",
                "operationId": "deleteComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du commentaire",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commentaire supprimé avec succès"
                    },
                    "404": {
                        "description": "Commentaire non trouvé"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}